itinerary CRUD methods sceleton added

+++ src/main/java/com/app/travel/repository/ItineraryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItineraryRepository extends JpaRepository<Itinerary, Long> {
+public interface ItineraryRepository extends JpaRepository<Itinerary, Integer> {
 }
